[chore]  RecordInfoWithBookmark에 of 함수 적용 (#57)
