--- src/main/java/org/recordy/server/record/controller/dto/response/RecordInfoWithBookmark.java
@@ -1,8 +1,0 @@
-package org.recordy.server.record.controller.dto.response;
-
-public record RecordInfoWithBookmark (
-        RecordInfo recordInfo,
-        Boolean isBookmark
-){
-
-}
