+++ src/main/java/org/recordy/server/record_stat/service/RecordStatService.java
@@ -1,7 +1,7 @@
 package org.recordy.server.record_stat.service;
 
 import java.util.List;
-import org.recordy.server.record.controller.dto.response.RecordInfoWithBookmark;
+import org.recordy.server.record.domain.usecase.RecordInfoWithBookmark;
 import org.recordy.server.record.domain.Record;
 import org.recordy.server.record_stat.domain.Bookmark;
 import org.recordy.server.record_stat.domain.usecase.Preference;
