+++ src/main/java/org/recordy/server/record_stat/service/impl/RecordStatServiceImpl.java
@@ -4,8 +4,8 @@ import java.util.List;
 import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.recordy.server.common.message.ErrorMessage;
-import org.recordy.server.record.controller.dto.response.RecordInfo;
-import org.recordy.server.record.controller.dto.response.RecordInfoWithBookmark;
+import org.recordy.server.record.domain.usecase.RecordInfo;
+import org.recordy.server.record.domain.usecase.RecordInfoWithBookmark;
 import org.recordy.server.record.domain.Record;
 import org.recordy.server.record.exception.RecordException;
 import org.recordy.server.record.repository.RecordRepository;
